type düzenlendi

+++ ArdRehber/ArdRehber/Controllers/DepartmentsController.cs
@@ -12,13 +12,14 @@ using ArdRehber.Dtos;
 using Microsoft.AspNetCore.Authorization;
 using ArdRehber.FluentValidation;
 using FluentValidation.Results;
+using ArdRehber.Enums;
 
 namespace ArdRehber.Controllers
 {
     [Authorize]
     [Route("api/[controller]")]
     [ApiController]
-    public class DepartmentsController : ControllerBase
+    public class DepartmentsController : BaseController
     {
         private readonly DataContext _context;
 
@@ -65,21 +66,39 @@ namespace ArdRehber.Controllers
         [HttpGet]
         public async Task<ActionResult<IEnumerable<Department>>> GetDepartments()
         {
-            return await _context.Departments.ToListAsync(); // INCLUDE BURADA YAPILDI.    Include(x=>x.Persons).
+            EUserType userTyp = GetUserType();
+            if (userTyp == EUserType.Admin)
+            {
+                return await _context.Departments.ToListAsync(); // INCLUDE BURADA YAPILDI.    Include(x=>x.Persons).
+            }
+            else
+            {
+                return BadRequest("Bu alana erişim yetkiniz bulunmamaktadır.");
+            }
+
         }
 
         // GET: api/Departments/5
         [HttpGet("{id}")]
         public async Task<ActionResult<Department>> GetDepartment(int id)
         {
-            var department = await _context.Departments.FindAsync(id);
-
-            if (department == null)
+            EUserType userTyp = GetUserType();
+            if (userTyp == EUserType.Admin)
             {
-                return NotFound();
+                var department = await _context.Departments.FindAsync(id);
+
+                if (department == null)
+                {
+                    return NotFound();
+                }
+
+                return department;
+            }
+            else
+            {
+                return BadRequest("Bu alana erişim yetkiniz bulunmamaktadır.");
             }
 
-            return department;
         }
 
         
@@ -88,55 +107,72 @@ namespace ArdRehber.Controllers
         [HttpPost]
         public async Task<ActionResult<Department>> PostDepartment(DepartmentDto departmentDto)
         {
-
-            DepartmentValidator departmentValidator = new DepartmentValidator();
-            ValidationResult results = departmentValidator.Validate(departmentDto);
-
-            if (results.IsValid == false)
+            EUserType userTyp = GetUserType();
+            if (userTyp == EUserType.Admin)
             {
-                return BadRequest(results.Errors[0].ErrorMessage);
-            }
+                DepartmentValidator departmentValidator = new DepartmentValidator();
+                ValidationResult results = departmentValidator.Validate(departmentDto);
 
-            if (departmentDto.DepartmentId > 0)
-            {
-                var department = this.UpdateDepartment(departmentDto).Result;
-               
+                if (results.IsValid == false)
+                {
+                    return BadRequest(results.Errors[0].ErrorMessage);
+                }
 
+                if (departmentDto.DepartmentId > 0)
+                {
+                    var department = this.UpdateDepartment(departmentDto).Result;
+
+
+                }
+                else
+                {
+                    var department = this.AddDepartment(departmentDto).Result;
+                    if (department == null)
+                    {
+                        return BadRequest("Böyle bir departman zaten var.");
+                    }
+
+                    departmentDto.DepartmentId = department.DepartmentId;
+                }
+
+                return Ok(departmentDto);
+
+
+
+                //_context.Departments.Add(department);
+                //await _context.SaveChangesAsync();
+                //return CreatedAtAction("GetDepartment", new { id = department.DepartmentId }, department);
             }
             else
             {
-                var department = this.AddDepartment(departmentDto).Result;
-                if (department == null)
-                {
-                    return BadRequest("Böyle bir departman zaten var.");
-                }
-
-                departmentDto.DepartmentId = department.DepartmentId;
+                return BadRequest("Bu alana erişim yetkiniz bulunmamaktadır.");
             }
-                      
-            return Ok(departmentDto);
 
-
-
-            //_context.Departments.Add(department);
-            //await _context.SaveChangesAsync();
-            //return CreatedAtAction("GetDepartment", new { id = department.DepartmentId }, department);
         }
 
         // DELETE: api/Departments/5
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteDepartment(int id)
         {
-            var department = await _context.Departments.FindAsync(id);
-            if (department == null)
+            EUserType userTyp = GetUserType();
+            if (userTyp == EUserType.Admin)
             {
-                return NotFound();
+                var department = await _context.Departments.FindAsync(id);
+                if (department == null)
+                {
+                    return NotFound();
+                }
+
+                _context.Departments.Remove(department);
+                await _context.SaveChangesAsync();
+
+                return NoContent();
+            }
+            else
+            {
+                return BadRequest("Bu alana erişim yetkiniz bulunmamaktadır.");
             }
 
-            _context.Departments.Remove(department);
-            await _context.SaveChangesAsync();
-
-            return NoContent();
         }
 
         private bool DepartmentExists(int id)
